Fixed separate component for displaying assigned members, fixed som styling related to members in task, listPage and popup as well as the dropdown menu for moving tasks.
