+++ src/styling/ListPage.module.css
@@ -1,16 +1,23 @@
-.listContainer{
+.listContainer {
   height: 80vh;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 
-h2{
+h2 {
   margin-bottom: 1em;
   text-align: center;
 }
 
-.listItem{
-  padding: .7em;
+.listItem {
+  width: 40%;
+  height: 5em;
+  padding: 0 0.7em;
   border: 2px solid red;
   display: flex;
+  align-items: center;
   justify-content: space-between;
   margin: 0.5em 1em 15px 0.5em;
   border-radius: 25px 25px;
