+++ src/styling/TaskPopup.module.css
@@ -82,33 +82,11 @@ h4 {
   outline: 2px solid rgb(182, 182, 182);
 }
 
-/***** Assigned Users ******/
 .assignedUsers_div {
   width: fit-content;
   position: relative;
-  padding: 0.5em;
   border-radius: 10px;
   margin-bottom: 1.5em;
-}
-
-.displayMembers_div {
-  margin-top: 0.5em;
-  display: flex;
-  align-items: center;
-  gap: 0.7em;
-}
-
-.renderedUser {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.userAvatar {
-  width: 30px;
-  height: 30px;
-  border-radius: 50%;
 }
 
 .MainField {
